Fix foreign_key error and rename models
